Navigation implemented and separate styles file
