+++ App.tsx
@@ -1,29 +1,18 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
+import 'react-native-gesture-handler';
 import React from 'react';
-import { SafeAreaView, StyleSheet, Dimensions } from 'react-native';
+import { SafeAreaView, StyleSheet, Dimensions, View } from 'react-native';
 import SignUpScreen from './src/screens/SignUpScreen';
+import { NavigationContainer } from '@react-navigation/native';
+import { StackNavigator } from './src/navigation/StackNavigator';
 
 function App(): JSX.Element {
   return (
-    <SafeAreaView style={styles.mainContainer}>
-      <SignUpScreen />
-    </SafeAreaView>
+
+    <NavigationContainer>
+        <StackNavigator />
+      </NavigationContainer>
   );
 }
-const styles = StyleSheet.create({
-  mainContainer: {
-    justifyContent: 'center',
-    alignSelf: 'center',
-    width: '100%',
-    maxWidth: Dimensions.get('window').width * 0.90,
-    marginHorizontal: 'auto',
-  },
-});
+
 
 export default App;
