+++ package.json
@@ -11,8 +11,14 @@
   },
   "dependencies": {
     "@react-native-community/checkbox": "^0.5.15",
+    "@react-native-masked-view/masked-view": "^0.2.9",
+    "@react-navigation/native": "^6.1.7",
+    "@react-navigation/stack": "^6.3.17",
     "react": "18.2.0",
     "react-native": "0.71.11",
+    "react-native-gesture-handler": "^2.12.0",
+    "react-native-safe-area-context": "^4.6.3",
+    "react-native-screens": "^3.22.0",
     "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
