+++ src/components/ButtonPrimary.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Image, StyleSheet, Text, TouchableOpacity } from 'react-native';
 import { ButtonProps } from '../interfaces/interfaces';
+import styles from '../appTheme/AppTheme';
 
 
 export const ButtonPrimary = ({ title, onPress, imgSource, isValid }: ButtonProps) => {
@@ -17,42 +18,3 @@ export const ButtonPrimary = ({ title, onPress, imgSource, isValid }: ButtonProp
     );
 };
 
-const styles = StyleSheet.create({
-    buttonContainer: {
-        alignItems: 'center',
-        alignSelf: 'center',
-        borderRadius: 10,
-        marginTop: 10,
-        paddingVertical: 10,
-        paddingHorizontal: 20,
-        width: '95%',
-        height: 50,
-        flexDirection: 'row',
-        justifyContent: 'center',
-        shadowColor: '#000',
-        shadowOffset: {
-            width: 0,
-            height: 4,
-        },
-        shadowOpacity: 0.30,
-        shadowRadius: 3,
-        elevation: 4,
-    },
-    primaryButtonValid: {
-        backgroundColor: '#2774D5',
-    },
-    primaryButtonInvalid: {
-        backgroundColor: '#A0A0A0',
-    },
-    buttonText: {
-        color: 'white',
-        fontSize: 18,
-        fontWeight: 'bold',
-    },
-    buttonImage: {
-        position: 'absolute',
-        left: '10%',
-        width: 25,
-        height: 25,
-    },
-});
