+++ src/screens/LogInScreen.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+
+export const LogInScreen = () => {
+  return (
+    <View>
+        <Text>Log In Screen</Text>
+    </View>
+  )
+}
